supabase/functions/ML-prediction/requirements.txt
fastapi==0.103.1
uvicorn[standard]==0.23.2
pydantic==2.4.1
supabase==1.0.3
spacy==3.6.1
scikit-learn==1.3.0
pandas==2.1.0
python-dotenv==1.0.0
python-multipart==0.0.6
